rcu: Move RCU-related source code to kernel/rcu directory
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
Git-commit: 4102adab9189c8ea2f0cdd2f88345fd25d2790f1
[pdaly@codeaurora.org: resolve trivial merge conflicts]
(cherry picked from commit 36904ca759039f4c5c1d4e1ad56ee9d1cb8ac699)
(cherry picked from commit 4c3b0a53e0379d7ec079d42aa9b5034b23d17541)
